[#34] Split up and update tests
